Rename page_layout to layout.
