+++ src/BUILD.bazel
@@ -12,9 +12,9 @@ filegroup(
         "builtin_modules.rs",
         "code_blocks.rs",
         "dom_helpers.rs",
+        "layout.rs",
         "modules.rs",
         "notebook.rs",
-        "page_layout.rs",
         "prism.rs",
     ],
     visibility = ["//:__pkg__"],
